--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,6 @@
 module github.com/gogo/protobuf
 
-require github.com/kisielk/errcheck v1.1.0 // indirect
+require (
+	github.com/kisielk/errcheck v1.2.0 // indirect
+	github.com/kisielk/gotool v1.0.0 // indirect
+)