--- a/protoc-gen-gogo/generator/generator.go
+++ b/protoc-gen-gogo/generator/generator.go
@@ -3148,7 +3148,7 @@ func (g *Generator) generateCommonMethods(mc *msgCtx) {
 		gogoproto.IsUnsafeMarshaler(g.file.FileDescriptorProto, mc.message.DescriptorProto) {
 		if gogoproto.IsStableMarshaler(g.file.FileDescriptorProto, mc.message.DescriptorProto) {
 			g.P("b = b[:cap(b)]")
-			g.P("n, err := m.MarshalTo(b)")
+			g.P("n, err := m.MarshalToSizedBuffer(b)")
 			g.P("if err != nil {")
 			g.In()
 			g.P("return nil, err")
@@ -3162,7 +3162,7 @@ func (g *Generator) generateCommonMethods(mc *msgCtx) {
 			g.P("} else {")
 			g.In()
 			g.P("b = b[:cap(b)]")
-			g.P("n, err := m.MarshalTo(b)")
+			g.P("n, err := m.MarshalToSizedBuffer(b)")
 			g.P("if err != nil {")
 			g.In()
 			g.P("return nil, err")