… go.sum
… test/casttype/combos/both/casttype.pb.go
… test/casttype/combos/marshaler/casttype.pb.go
… test/casttype/combos/neither/casttype.pb.go
… test/casttype/combos/unmarshaler/casttype.pb.go
… test/castvalue/combos/both/castvalue.pb.go
… test/castvalue/combos/marshaler/castvalue.pb.go
… test/combos/both/thetest.pb.go
… test/combos/marshaler/thetest.pb.go
… test/combos/unmarshaler/thetest.pb.go
… test/data/data.pb.go
… test/deterministic/deterministic.pb.go
… test/enumdecl/enumdecl.pb.go
… test/enumdecl_all/enumdeclall.pb.go
… test/enumstringer/enumstringer.pb.go
… test/example/example.pb.go
… test/filedotname/file.dot.pb.go
… test/fuzztests/fuzz.pb.go
… test/group/group.pb.go
… test/importcustom-issue389/imported/a.pb.go
… test/importcustom-issue389/importing/c.pb.go
… test/importduplicate/importduplicate.pb.go
… test/indeximport-issue72/index/index.pb.go
… test/indeximport-issue72/indeximport.pb.go
… test/int64support/object.pb.go
… test/issue260/issue260.pb.go
… test/issue261/issue261.pb.go
… test/issue262/timefail.pb.go
… test/issue312/events/events.pb.go
… test/issue322/issue322.pb.go