.)(eziS gnitupmoc-erp )yledetaeper dna( ylevisrucer diova ot ,)elbissop nehw( drawkcab dezilaires era stcejbo taht os cigol eht lla esreveR

Fixes #559