@apelisse 2019-04-11
Not sure why the tests are failing, seems unrelated to the change, tests are passing.