@jmarais 2019-04-12
I updated master to use go12.4 and protoc3.7.1. However your current travis build still has the old travis config. I am not sure how to trigger a new build. Can you make another commit to (hopefully) trigger a new build with the new master .travis config?