@apelisse 2019-04-12
I’ll do that first time tomorrow and address Jordan’s comments, thanks

On Thu, Apr 11, 2019 at 10:53 PM jmarais <notifications@github.com> wrote:

> I updated master to use go12.4 and protoc3.7.1. However your current
> travis build still has the old travis config. I am not sure how to trigger
> a new build. Can you make another commit to (hopefully) trigger a new build
> with the new master .travis config?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/gogo/protobuf/pull/560#issuecomment-482447120>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAOXDLHVYHZDH3UGLLL23F3PQAPWRANCNFSM4HFLVXZA>
> .
>