@jmarais 2019-04-12
[We can perhaps add the reverse marshaling as a message/file option. ](https://github.com/gogo/protobuf/issues/559#issuecomment-481811977)

I would guess reverse marshalling changes the resulting marshalled bytes.  If that is the case I think for the short term adding this as an option if a safer bet. I would rather want to add it as another [plugin](https://github.com/gogo/protobuf/tree/master/plugin) and give other users the option to switch to it when appropriate.

If this is added as a file/message option, you can then leverage the a vanity binary in order to select the reverse-marshal plugin on generation time. 
In the [vanity example](https://github.com/gogo/protobuf/blob/master/protoc-gen-gogofast/main.go#L41) you can create an option to ```TurnOnReverseMarshalerAll```.

What do you think?