@liggitt 2019-04-12
> I would guess reverse marshalling changes the resulting marshalled bytes. 

It definitely shouldn't. If it does, we can't use this solution. We need sufficient before/after fixture testing to prove there is no visible change.