@apelisse 2019-04-12
> > I would guess reverse marshalling changes the resulting marshalled bytes.
>
> It definitely shouldn't. If it does, we can't use this solution. We need sufficient before/after fixture testing to > prove there is no visible change.

Agree, it shouldn't. We should build confidence that this code doesn't make performance worse on other paths, and also verify that the generated bytes are the same. Jenny and I will be working on this today (and trying some integration of this change with Kubernetes, that would give us some more confidence too).

> If that is the case I think for the short term adding this as an option if a safer bet.

My main problem with the option/plugin route is that we will have to keep both versions of the code, and I'm a little concerned that we end-up having a drift between the two. I don't see how we can maintain this without duplicating, it's too different. 

I think we need confidence that the change is an actual no-op, we can maybe get that confidence by having more fuzzy-testing, comparing the protobuf bytes rather than the marshal/unmarshal round-trip.

> Can you make another commit to (hopefully) trigger a new build with the new master .travis config?

I just rebased and pushed, tests are running now!