@apelisse 2019-04-12
> We're trying to run this code with Kubernetes, we have at least one bug (that wasn't caught by the test-suite here), we're trying to identify it and then we'll add a test.

False alarm, it's because of things we do wrong in Kubernetes. I'm able to create a cluster, connect, etc... Things seem to be working.