@smarterclayton 2019-04-14
Are there benchmark numbers for this for both small and large objects?  Would like to see benchcmp here