@jmarais 2019-04-15
> My main problem with the option/plugin route is that we will have to keep both versions of the code, and I'm a little concerned that we end-up having a drift between the two. I don't see how we can maintain this without duplicating, it's too different.

That is a good point. However, as you mentioned more confidence would be nice. I will see if I can run this against our internal regression tests today. 

> Are there benchmark numbers for this for both small and large objects? Would like to see benchcmp here

More benchmarks would be nice. We do have ```make bench``` in the repo. However it would be nice to have explicit benchmarks for a range of object sizes like the initial benchmarks submitted with the case [#559](https://github.com/gogo/protobuf/issues/559#issue-430756919)