@jmarais 2019-04-15
```bash
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkNidOptNativeProtoMarshal-5                                       486           656           +34.98%
BenchmarkNinOptNativeProtoMarshal-5                                       457           635           +38.95%
BenchmarkNidRepNativeProtoMarshal-5                                       1726          2508          +45.31%
BenchmarkNinRepNativeProtoMarshal-5                                       1699          2604          +53.27%
BenchmarkNidRepPackedNativeProtoMarshal-5                                 1636          2086          +27.51%
BenchmarkNinRepPackedNativeProtoMarshal-5                                 1614          2083          +29.06%
BenchmarkNidOptStructProtoMarshal-5                                       1771          3241          +83.00%
BenchmarkNinOptStructProtoMarshal-5                                       1613          2062          +27.84%
BenchmarkNidRepStructProtoMarshal-5                                       4109          7204          +75.32%
BenchmarkNinRepStructProtoMarshal-5                                       3781          5026          +32.93%
BenchmarkNidEmbeddedStructProtoMarshal-5                                  1058          1631          +54.16%
BenchmarkNinEmbeddedStructProtoMarshal-5                                  1003          1209          +20.54%
BenchmarkNidNestedStructProtoMarshal-5                                    12325         20744         +68.31%
BenchmarkNinNestedStructProtoMarshal-5                                    10130         10791         +6.53%
BenchmarkNidOptCustomProtoMarshal-5                                       188           391           +107.98%
BenchmarkCustomDashProtoMarshal-5                                         184           213           +15.76%
BenchmarkNinOptCustomProtoMarshal-5                                       215           242           +12.56%
BenchmarkNidRepCustomProtoMarshal-5                                       449           1043          +132.29%
BenchmarkNinRepCustomProtoMarshal-5                                       458           1042          +127.51%
BenchmarkNinOptNativeUnionProtoMarshal-5                                  129           140           +8.53%
BenchmarkNinOptStructUnionProtoMarshal-5                                  253           272           +7.51%
BenchmarkNinEmbeddedStructUnionProtoMarshal-5                             435           534           +22.76%
BenchmarkNinNestedStructUnionProtoMarshal-5                               358           393           +9.78%
BenchmarkTreeProtoMarshal-5                                               308           369           +19.81%
BenchmarkOrBranchProtoMarshal-5                                           635           1055          +66.14%
BenchmarkAndBranchProtoMarshal-5                                          652           1073          +64.57%
BenchmarkLeafProtoMarshal-5                                               226           278           +23.01%
BenchmarkDeepTreeProtoMarshal-5                                           460           666           +44.78%
BenchmarkADeepBranchProtoMarshal-5                                        599           973           +62.44%
BenchmarkAndDeepBranchProtoMarshal-5                                      1082          1885          +74.21%
BenchmarkDeepLeafProtoMarshal-5                                           380           596           +56.84%
BenchmarkNilProtoMarshal-5                                                99.8          99.4          -0.40%
BenchmarkNidOptEnumProtoMarshal-5                                         119           130           +9.24%
BenchmarkNinOptEnumProtoMarshal-5                                         159           180           +13.21%
BenchmarkNidRepEnumProtoMarshal-5                                         319           427           +33.86%
BenchmarkNinRepEnumProtoMarshal-5                                         319           422           +32.29%
BenchmarkNinOptEnumDefaultProtoMarshal-5                                  160           171           +6.88%
BenchmarkAnotherNinOptEnumProtoMarshal-5                                  159           180           +13.21%
BenchmarkAnotherNinOptEnumDefaultProtoMarshal-5                           161           179           +11.18%
BenchmarkTimerProtoMarshal-5                                              189           208           +10.05%
BenchmarkMyExtendableProtoMarshal-5                                       509           798           +56.78%
BenchmarkOtherExtenableProtoMarshal-5                                     1155          1647          +42.60%
BenchmarkNestedDefinitionProtoMarshal-5                                   449           579           +28.95%
BenchmarkNestedDefinition_NestedMessageProtoMarshal-5                     249           291           +16.87%
BenchmarkNestedDefinition_NestedMessage_NestedNestedMsgProtoMarshal-5     185           198           +7.03%
BenchmarkNestedScopeProtoMarshal-5                                        420           511           +21.67%
BenchmarkNinOptNativeDefaultProtoMarshal-5                                443           650           +46.73%
BenchmarkCustomContainerProtoMarshal-5                                    255           660           +158.82%
BenchmarkCustomNameNidOptNativeProtoMarshal-5                             478           653           +36.61%
BenchmarkCustomNameNinOptNativeProtoMarshal-5                             448           659           +47.10%
BenchmarkCustomNameNinRepNativeProtoMarshal-5                             1711          2589          +51.32%
BenchmarkCustomNameNinStructProtoMarshal-5                                2152          2737          +27.18%
BenchmarkCustomNameCustomTypeProtoMarshal-5                               467           1205          +158.03%
BenchmarkCustomNameNinEmbeddedStructUnionProtoMarshal-5                   423           538           +27.19%
BenchmarkCustomNameEnumProtoMarshal-5                                     214           238           +11.21%
BenchmarkNoExtensionsMapProtoMarshal-5                                    224           270           +20.54%
BenchmarkUnrecognizedProtoMarshal-5                                       133           140           +5.26%
BenchmarkUnrecognizedWithInnerProtoMarshal-5                              313           369           +17.89%
BenchmarkUnrecognizedWithInner_InnerProtoMarshal-5                        86.0          96.5          +12.21%
BenchmarkUnrecognizedWithEmbedProtoMarshal-5                              238           407           +71.01%
BenchmarkUnrecognizedWithEmbed_EmbeddedProtoMarshal-5                     78.9          94.1          +19.26%
BenchmarkNodeProtoMarshal-5                                               215           244           +13.49%
BenchmarkNonByteCustomTypeProtoMarshal-5                                  200           279           +39.50%
BenchmarkNidOptNonByteCustomTypeProtoMarshal-5                            179           370           +106.70%
BenchmarkNinOptNonByteCustomTypeProtoMarshal-5                            227           278           +22.47%
BenchmarkNidRepNonByteCustomTypeProtoMarshal-5                            539           978           +81.45%
BenchmarkNinRepNonByteCustomTypeProtoMarshal-5                            519           983           +89.40%
BenchmarkProtoTypeProtoMarshal-5                                          184           203           +10.33%
```

This is the current compare between ```master``` and ```OtLahsram```