@apelisse 2019-04-17
OK We've made a lot of progress. We've removed the interface and don't check against it, which makes it much easier to always call `LahsramOt` properly (and avoiding going back to doing a lot of `Size()` if the test fails because of pointers or what-not), and it also removes the extra condition. `XXX_Marshal` still calls `Size()` and then `MarshalTo()` which leads to a double call to `Size()`, that doesn't sit necessarily very well with micro-benchmarks like the ones below, but I would expect performance on real-life objects, especially deeply nested ones, to be much better (as we can see by looking at the not very deep "NestedStruct" tests below, which is still doing -13%).

`BenchmarkMyExtendableProtoMarshal` is doing terrible, and I have really no idea why, since it's just supposed to copy bits the exact same way ...

```
benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkNidOptNativeProtoMarshal-12                                       555           495           -10.81%
BenchmarkNinOptNativeProtoMarshal-12                                       521           514           -1.34%
BenchmarkNidRepNativeProtoMarshal-12                                       1507          1601          +6.24%
BenchmarkNinRepNativeProtoMarshal-12                                       1448          1646          +13.67%
BenchmarkNidRepPackedNativeProtoMarshal-12                                 1623          1622          -0.06%
BenchmarkNinRepPackedNativeProtoMarshal-12                                 1617          1537          -4.95%
BenchmarkNidOptStructProtoMarshal-12                                       1498          1348          -10.01%
BenchmarkNinOptStructProtoMarshal-12                                       1370          1267          -7.52%
BenchmarkNidRepStructProtoMarshal-12                                       2946          2936          -0.34%
BenchmarkNinRepStructProtoMarshal-12                                       2686          2631          -2.05%
BenchmarkNidEmbeddedStructProtoMarshal-12                                  1002          953           -4.89%
BenchmarkNinEmbeddedStructProtoMarshal-12                                  966           896           -7.25%
BenchmarkNidNestedStructProtoMarshal-12                                    7617          6642          -12.80%
BenchmarkNinNestedStructProtoMarshal-12                                    6585          5712          -13.26%
BenchmarkNidOptCustomProtoMarshal-12                                       230           256           +11.30%
BenchmarkCustomDashProtoMarshal-12                                         212           228           +7.55%
BenchmarkNinOptCustomProtoMarshal-12                                       217           257           +18.43%
BenchmarkNidRepCustomProtoMarshal-12                                       481           565           +17.46%
BenchmarkNinRepCustomProtoMarshal-12                                       509           548           +7.66%
BenchmarkNinOptNativeUnionProtoMarshal-12                                  131           140           +6.87%
BenchmarkNinOptStructUnionProtoMarshal-12                                  265           273           +3.02%
BenchmarkNinEmbeddedStructUnionProtoMarshal-12                             440           473           +7.50%
BenchmarkNinNestedStructUnionProtoMarshal-12                               345           335           -2.90%
BenchmarkTreeProtoMarshal-12                                               319           330           +3.45%
BenchmarkOrBranchProtoMarshal-12                                           606           608           +0.33%
BenchmarkAndBranchProtoMarshal-12                                          617           597           -3.24%
BenchmarkLeafProtoMarshal-12                                               255           288           +12.94%
BenchmarkDeepTreeProtoMarshal-12                                           452           473           +4.65%
BenchmarkADeepBranchProtoMarshal-12                                        552           537           -2.72%
BenchmarkAndDeepBranchProtoMarshal-12                                      901           798           -11.43%
BenchmarkDeepLeafProtoMarshal-12                                           417           437           +4.80%
BenchmarkNilProtoMarshal-12                                                121           126           +4.13%
BenchmarkNidOptEnumProtoMarshal-12                                         149           161           +8.05%
BenchmarkNinOptEnumProtoMarshal-12                                         183           212           +15.85%
BenchmarkNidRepEnumProtoMarshal-12                                         338           428           +26.63%
BenchmarkNinRepEnumProtoMarshal-12                                         355           415           +16.90%
BenchmarkNinOptEnumDefaultProtoMarshal-12                                  183           219           +19.67%
BenchmarkAnotherNinOptEnumProtoMarshal-12                                  173           219           +26.59%
BenchmarkAnotherNinOptEnumDefaultProtoMarshal-12                           168           228           +35.71%
BenchmarkTimerProtoMarshal-12                                              226           251           +11.06%
BenchmarkMyExtendableProtoMarshal-12                                       426           637           +49.53%
BenchmarkOtherExtenableProtoMarshal-12                                     852           1092          +28.17%
BenchmarkNestedDefinitionProtoMarshal-12                                   528           546           +3.41%
BenchmarkNestedDefinition_NestedMessageProtoMarshal-12                     302           332           +9.93%
BenchmarkNestedDefinition_NestedMessage_NestedNestedMsgProtoMarshal-12     207           230           +11.11%
BenchmarkNestedScopeProtoMarshal-12                                        513           508           -0.97%
BenchmarkNinOptNativeDefaultProtoMarshal-12                                544           552           +1.47%
BenchmarkCustomContainerProtoMarshal-12                                    318           349           +9.75%
BenchmarkCustomNameNidOptNativeProtoMarshal-12                             538           539           +0.19%
BenchmarkCustomNameNinOptNativeProtoMarshal-12                             567           518           -8.64%
BenchmarkCustomNameNinRepNativeProtoMarshal-12                             1538          1632          +6.11%
BenchmarkCustomNameNinStructProtoMarshal-12                                1744          1590          -8.83%
BenchmarkCustomNameCustomTypeProtoMarshal-12                               566           603           +6.54%
BenchmarkCustomNameNinEmbeddedStructUnionProtoMarshal-12                   492           466           -5.28%
BenchmarkCustomNameEnumProtoMarshal-12                                     251           267           +6.37%
BenchmarkNoExtensionsMapProtoMarshal-12                                    262           262           +0.00%
BenchmarkUnrecognizedProtoMarshal-12                                       154           168           +9.09%
BenchmarkUnrecognizedWithInnerProtoMarshal-12                              344           373           +8.43%
BenchmarkUnrecognizedWithInner_InnerProtoMarshal-12                        63.6          69.8          +9.75%
BenchmarkUnrecognizedWithEmbedProtoMarshal-12                              282           302           +7.09%
BenchmarkUnrecognizedWithEmbed_EmbeddedProtoMarshal-12                     64.7          70.5          +8.96%
BenchmarkNodeProtoMarshal-12                                               256           293           +14.45%
BenchmarkNonByteCustomTypeProtoMarshal-12                                  271           296           +9.23%
BenchmarkNidOptNonByteCustomTypeProtoMarshal-12                            294           329           +11.90%
BenchmarkNinOptNonByteCustomTypeProtoMarshal-12                            281           304           +8.19%
BenchmarkNidRepNonByteCustomTypeProtoMarshal-12                            657           696           +5.94%
BenchmarkNinRepNonByteCustomTypeProtoMarshal-12                            631           679           +7.61%
BenchmarkProtoTypeProtoMarshal-12                                          223           233           +4.48%
```