@apelisse 2019-07-10
Rebased on top of master.

I've also added a new test that tries against a recursive map. (I had to change a random number somewhere to generate bigger maps):
```
benchmark                             old ns/op     new ns/op     delta
BenchmarkDeepMapProtoMarshal-12       7978          2263          -71.63%
```

I think this is ready to merge. We have tested with Kubernetes without any backward compatibility issues or problems whatsoever.